+++ src/main/java/br/org/oficinadasmeninas/domain/donation/dto/DonationDto.java
@@ -5,6 +5,6 @@ import java.util.UUID;
 
 import br.org.oficinadasmeninas.domain.donation.DonationStatusEnum;
 
-public record DonationDto(UUID id, long value, LocalDateTime donationAt, UUID userId, DonationStatusEnum status) {
+public record DonationDto(UUID id, double value, LocalDateTime donationAt, UUID userId, DonationStatusEnum status) {
 
 }
