+++ src/main/java/br/org/oficinadasmeninas/domain/donation/dto/DonationItemDto.java
@@ -2,6 +2,6 @@ package br.org.oficinadasmeninas.domain.donation.dto;
 
 import java.util.Optional;
 
-public record DonationItemDto(Long value, boolean isRecurring, Optional<Integer> cycles) {
+public record DonationItemDto(double value, boolean isRecurring, Optional<Integer> cycles) {
 
 }
